pyproject.toml
[project]
name = "tkcalc"
version = "0.9.3"
description = "Калькулятор Теплокомплект"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycairo>=1.28.0",
    "pygobject>=3.52.3",
]

[dependency-groups]
dev = [
    "nuitka>=2.7.3",
    "ruff>=0.11.11",
]
